+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changes to Calva.
 ## [Unreleased]
 - Fix: [debugging stops working](https://github.com/BetterThanTomorrow/calva/issues/3224)
 
+- [Dragging sexps with adjacent comments](https://github.com/BetterThanTomorrow/calva/issues/3073)
+
 ## [2.0.588] - 2026-05-15
 
 - [Make API clojure editing independent of a `TextEditor` instance](https://github.com/BetterThanTomorrow/calva/issues/3220)